+++ hooks/use-color-scheme.web.ts
@@ -1,6 +1,8 @@
 import { useEffect, useState } from 'react';
 import { useColorScheme as useRNColorScheme } from 'react-native';
 
+// ----------------------------------------------------------------------
+
 /**
  * To support static rendering, this value needs to be re-calculated on the client side for web
  */
